+++ packages/example/src/core/user/model.ts
@@ -4,18 +4,21 @@ import { model } from '@mondrian-framework/model'
 
 export type User = model.Infer<typeof User>
 export const User = () =>
-  model.entity({
-    id: idType,
-    firstName: model.string(),
-    lastName: model.string(),
-    email: model.email(),
-    posts: model.array(Post),
-    givenLikes: model.array(Like),
-    followers: model.array(Follower),
-    followeds: model.array(Follower),
-    registeredAt: model.datetime(),
-    loginAt: model.datetime(),
-  })
+  model.entity(
+    {
+      id: idType,
+      firstName: model.string(),
+      lastName: model.string(),
+      email: model.email(),
+      posts: model.array(Post),
+      givenLikes: model.array(Like),
+      followers: model.describe(model.array(Follower), 'Users that follows me'),
+      followeds: model.describe(model.array(Follower), 'Users followed by me'),
+      registeredAt: model.datetime(),
+      loginAt: model.datetime(),
+    },
+    { description: 'User of the system' },
+  )
 
 export const MyUser = () => User
 export type MyUser = User