+++ packages/middleware/rate-limiter/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @mondrian-framework/rate-limiter
 
+## 2.0.13
+
+### Patch Changes
+
+- Various improvements
+- Updated dependencies
+  - @mondrian-framework/model@2.0.13
+  - @mondrian-framework/module@2.0.13
+  - @mondrian-framework/utils@2.0.13
+
 ## 2.0.12
 
 ### Patch Changes