+++ packages/model/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @mondrian-framework/model
 
+## 2.0.13
+
+### Patch Changes
+
+- Various improvements
+- Updated dependencies
+  - @mondrian-framework/utils@2.0.13
+
 ## 2.0.12
 
 ### Patch Changes